+++ pubnub/endpoints/membership/get_members.py
@@ -2,8 +2,9 @@ import six
 
 from pubnub import utils
 from pubnub.endpoints.endpoint import Endpoint
+from pubnub.managers import TokenManagerProperties
 from pubnub.models.consumer.membership import PNGetMembersResult
-from pubnub.enums import HttpMethod, PNOperationType
+from pubnub.enums import HttpMethod, PNOperationType, PNResourceType
 from pubnub.exceptions import PubNubException
 
 
@@ -98,3 +99,9 @@ class GetMembers(Endpoint):
 
     def name(self):
         return 'Get members'
+
+    def get_tms_properties(self):
+        return TokenManagerProperties(
+            resource_type=PNResourceType.SPACE,
+            resource_id=self._space_id if self._space_id is not None else ""
+        )