+++ pubnub/endpoints/space/delete_space.py
@@ -1,8 +1,9 @@
 import six
 
 from pubnub.endpoints.endpoint import Endpoint
+from pubnub.managers import TokenManagerProperties
 from pubnub.models.consumer.space import PNDeleteSpaceResult
-from pubnub.enums import HttpMethod, PNOperationType
+from pubnub.enums import HttpMethod, PNOperationType, PNResourceType
 from pubnub.exceptions import PubNubException
 
 
@@ -52,3 +53,9 @@ class DeleteSpace(Endpoint):
 
     def name(self):
         return 'Delete space'
+
+    def get_tms_properties(self):
+        return TokenManagerProperties(
+            resource_type=PNResourceType.SPACE,
+            resource_id=self._space_id if self._space_id is not None else ""
+        )