+++ pubnub/endpoints/space/get_spaces.py
@@ -1,8 +1,9 @@
 import six
 
 from pubnub.endpoints.endpoint import Endpoint
+from pubnub.managers import TokenManagerProperties
 from pubnub.models.consumer.space import PNGetSpacesResult
-from pubnub.enums import HttpMethod, PNOperationType
+from pubnub.enums import HttpMethod, PNOperationType, PNResourceType
 
 
 class GetSpaces(Endpoint):
@@ -86,3 +87,9 @@ class GetSpaces(Endpoint):
 
     def name(self):
         return 'Get spaces'
+
+    def get_tms_properties(self):
+        return TokenManagerProperties(
+            resource_type=PNResourceType.SPACE,
+            resource_id=""
+        )