+++ pubnub/endpoints/space/update_space.py
@@ -2,8 +2,9 @@ import six
 
 from pubnub import utils
 from pubnub.endpoints.endpoint import Endpoint
+from pubnub.managers import TokenManagerProperties
 from pubnub.models.consumer.space import PNUpdateSpaceResult
-from pubnub.enums import HttpMethod, PNOperationType
+from pubnub.enums import HttpMethod, PNOperationType, PNResourceType
 from pubnub.exceptions import PubNubException
 
 
@@ -69,3 +70,9 @@ class UpdateSpace(Endpoint):
 
     def name(self):
         return 'Update space'
+
+    def get_tms_properties(self):
+        return TokenManagerProperties(
+            resource_type=PNResourceType.SPACE,
+            resource_id=self._space_id if self._space_id is not None else ""
+        )