+++ tests/functional/test_heartbeat.py
@@ -5,7 +5,6 @@ import json
 from pubnub.endpoints.presence.heartbeat import Heartbeat
 from pubnub.managers import TelemetryManager
 
-
 try:
     from mock import MagicMock
 except ImportError:
@@ -30,8 +29,8 @@ class TestHeartbeat(unittest.TestCase):
     def test_sub_single_channel(self):
         self.hb.channels('ch')
 
-        self.assertEquals(self.hb.build_path(), Heartbeat.HEARTBEAT_PATH
-                          % (pnconf.subscribe_key, 'ch'))
+        self.assertEqual(self.hb.build_path(), Heartbeat.HEARTBEAT_PATH
+                         % (pnconf.subscribe_key, 'ch'))
 
         self.assertEqual(self.hb.build_params_callback()({}), {
             'pnsdk': sdk_name,
@@ -44,8 +43,8 @@ class TestHeartbeat(unittest.TestCase):
     def test_hb_multiple_channels_using_list(self):
         self.hb.channels(['ch1', 'ch2', 'ch3'])
 
-        self.assertEquals(self.hb.build_path(), Heartbeat.HEARTBEAT_PATH
-                          % (pnconf.subscribe_key, "ch1,ch2,ch3"))
+        self.assertEqual(self.hb.build_path(), Heartbeat.HEARTBEAT_PATH
+                         % (pnconf.subscribe_key, "ch1,ch2,ch3"))
 
         self.assertEqual(self.hb.build_params_callback()({}), {
             'pnsdk': sdk_name,
@@ -58,8 +57,8 @@ class TestHeartbeat(unittest.TestCase):
     def test_hb_single_group(self):
         self.hb.channel_groups("gr")
 
-        self.assertEquals(self.hb.build_path(), Heartbeat.HEARTBEAT_PATH
-                          % (pnconf.subscribe_key, ","))
+        self.assertEqual(self.hb.build_path(), Heartbeat.HEARTBEAT_PATH
+                         % (pnconf.subscribe_key, ","))
 
         self.assertEqual(self.hb.build_params_callback()({}), {
             'channel-group': 'gr',
@@ -73,8 +72,8 @@ class TestHeartbeat(unittest.TestCase):
     def test_hb_multiple_groups_using_list(self):
         self.hb.channel_groups(['gr1', 'gr2', 'gr3'])
 
-        self.assertEquals(self.hb.build_path(), Heartbeat.HEARTBEAT_PATH
-                          % (pnconf.subscribe_key, ","))
+        self.assertEqual(self.hb.build_path(), Heartbeat.HEARTBEAT_PATH
+                         % (pnconf.subscribe_key, ","))
 
         self.assertEqual(self.hb.build_params_callback()({}), {
             'channel-group': 'gr1,gr2,gr3',
@@ -91,8 +90,8 @@ class TestHeartbeat(unittest.TestCase):
         state = {"name": "Alex", "count": 7}
         self.hb.channels('ch1,ch2').state(state)
 
-        self.assertEquals(self.hb.build_path(), Heartbeat.HEARTBEAT_PATH
-                          % (pnconf.subscribe_key, "ch1,ch2"))
+        self.assertEqual(self.hb.build_path(), Heartbeat.HEARTBEAT_PATH
+                         % (pnconf.subscribe_key, "ch1,ch2"))
 
         params = self.hb.build_params_callback()({})
         params['state'] = json.loads(six.moves.urllib.parse.unquote(params['state']))