+++ tests/functional/test_revoke.py
@@ -2,6 +2,7 @@ import unittest
 
 from pubnub import utils
 from pubnub.endpoints.access.revoke import Revoke
+from pubnub.enums import HttpMethod
 
 try:
     from mock import MagicMock
@@ -33,7 +34,7 @@ class TestRevoke(unittest.TestCase):
     def test_revoke_to_channel(self):
         self.revoke.channels('ch')
 
-        self.assertEquals(self.revoke.build_path(), Revoke.GRANT_PATH % pnconf.subscribe_key)
+        self.assertEqual(self.revoke.build_path(), Revoke.GRANT_PATH % pnconf.subscribe_key)
 
         pam_args = utils.prepare_pam_arguments({
             'timestamp': 123,
@@ -44,7 +45,10 @@ class TestRevoke(unittest.TestCase):
             'pnsdk': sdk_name,
             'uuid': self.pubnub.uuid
         })
-        sign_input = pnconf.subscribe_key + "\n" + pnconf.publish_key + "\n" + "grant\n" + pam_args
+        sign_input = HttpMethod.string(self.revoke.http_method()).upper() + "\n" + \
+            pnconf.publish_key + "\n" + \
+            self.revoke.build_path() + "\n" + \
+            pam_args + "\n"
         self.assertEqual(self.revoke.build_params_callback()({}), {
             'pnsdk': sdk_name,
             'uuid': self.pubnub.uuid,
@@ -53,7 +57,7 @@ class TestRevoke(unittest.TestCase):
             'r': '0',
             'w': '0',
             'm': '0',
-            'signature': utils.sign_sha256(pnconf.secret_key, sign_input)
+            'signature': "v2." + utils.sign_sha256(pnconf.secret_key, sign_input).rstrip("=")
         })
 
     def test_revoke_read_to_channel(self):
@@ -65,7 +69,7 @@ class TestRevoke(unittest.TestCase):
     def test_grant_read_and_write_to_channel_group(self):
         self.revoke.channel_groups(['gr1', 'gr2'])
 
-        self.assertEquals(self.revoke.build_path(), Revoke.GRANT_PATH % pnconf.subscribe_key)
+        self.assertEqual(self.revoke.build_path(), Revoke.GRANT_PATH % pnconf.subscribe_key)
 
         pam_args = utils.prepare_pam_arguments({
             'r': '0',
@@ -76,7 +80,10 @@ class TestRevoke(unittest.TestCase):
             'pnsdk': sdk_name,
             'uuid': self.pubnub.uuid
         })
-        sign_input = pnconf.subscribe_key + "\n" + pnconf.publish_key + "\n" + "grant\n" + pam_args
+        sign_input = HttpMethod.string(self.revoke.http_method()).upper() + "\n" + \
+            pnconf.publish_key + "\n" + \
+            self.revoke.build_path() + "\n" + \
+            pam_args + "\n"
         self.assertEqual(self.revoke.build_params_callback()({}), {
             'pnsdk': sdk_name,
             'uuid': self.pubnub.uuid,
@@ -85,5 +92,5 @@ class TestRevoke(unittest.TestCase):
             'm': '0',
             'timestamp': '123',
             'channel-group': 'gr1,gr2',
-            'signature': utils.sign_sha256(pnconf.secret_key, sign_input)
+            'signature': "v2." + utils.sign_sha256(pnconf.secret_key, sign_input).rstrip("=")
         })