+++ tests/functional/test_stringify.py
@@ -36,14 +36,16 @@ class TestStringify(unittest.TestCase):
         assert str(result) == "Group contains following channels: qwer, asdf, zxcv"
 
     def test_audit(self):
-        result = PNAccessManagerAuditResult(None, None, None, None, 3600, True, False, True)
+        result = PNAccessManagerAuditResult(None, None, None, None, 3600, True, False, True, False)
 
-        assert str(result) == "Current permissions are valid for 3600 minutes: read True, write False, manage: True"
+        assert str(result) == \
+            "Current permissions are valid for 3600 minutes: read True, write False, manage: True, delete: False"
 
     def test_grant(self):
-        result = PNAccessManagerGrantResult(None, None, None, None, 3600, True, False, True)
+        result = PNAccessManagerGrantResult(None, None, None, None, 3600, True, False, True, False)
 
-        assert str(result) == "New permissions are set for 3600 minutes: read True, write False, manage: True"
+        assert str(result) == \
+            "New permissions are set for 3600 minutes: read True, write False, manage: True, delete: False"
 
     def test_history(self):
         assert str(PNHistoryResult(None, 123, 789)) == "History result for range 123..789"